feat: sonyflake

+++ packages/server/application/service/admin/thread_comment_application_service.go
@@ -15,7 +15,7 @@ func NewThreadCommentApplicationService(threadCommentDatasource *datasource_admi
 
 type ThreadCommentApplicationServiceFindByIDParams struct {
 	Ctx       context.Context
-	CommentID int
+	CommentID uint64
 }
 
 func (svc *ThreadCommentApplicationService) Delete(params ThreadCommentApplicationServiceFindByIDParams) error {
